% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exteriors_study_curb_appeal}
\alias{exteriors_study_curb_appeal}
\title{Exteriors survey responses for 360 individuals, for evaluating new curb appeal words.}
\format{
A data frame with 125 rows and 5 variables:
\describe{
  \item{arm}{treatment arm, as factor}
  \item{pre_q}{pre-intervention response, lower-cased and lemmatized}
  \item{post_q}{post-intervention response, lower-cased and lemmatized}
  \item{housing_involved}{indicates whether respondent is involved in housing sales and development, as logical}
  \item{housing_purchased}{indicates whether respondent has tried to purchase a house in past ten years, as logical}
}
}
\usage{
exteriors_study_curb_appeal
}
\description{
A dataset containing exteriors survey responses of 357 individuals, with
89 responses per treatment arm. This dataset contains responses for the
following treatment arms:
    * strong architecture
    * curb appeal
    * control
}
\keyword{datasets}
