% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exteriors_study_large}
\alias{exteriors_study_large}
\title{500 exteriors survey responses, all treatment arms}
\format{
A data frame with 470 rows and 5 variables:
\describe{
\item{arm}{treatment arm, as factor}
\item{pre_q}{pre-intervention response, lower-cased and lemmatized}
\item{post_q}{post-intervention response, lower-cased and lemmatized}
\item{housing_involved}{indicates whether respondent is involved in housing sales and development, as logical}
\item{housing_purchased}{indicates whether respondent has tried to purchase a house in past ten years, as logical}
}
}
\usage{
exteriors_study_large
}
\description{
A dataset containing exteriors survey responses for almost 500 individuals,
with almost 100 responses per treatment arm. This dataset contains responses
for the following treatment arms:
\itemize{
\item strong architecture
\item weak architecture
\item graphical architecture
\item curb appeal
\item control
}
}
\keyword{datasets}
